Included variants that were skipped because of source indirection
The json source for many items (sword +1, etc) redirects part of its
stats because the weapon +1 could be any weapon, and the source used the
data to build these items in detail. I don't, so I just create a generic
item.

It means I miss some of the item filtering, because for some reason the
rarity and frequency weren't tied to the base magic item, but whatever.